LRQA-39804 Move testRelevantChange properties to base class
