+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/BatchTestClassGroup.java
@@ -16,6 +16,7 @@ package com.liferay.jenkins.results.parser;
 
 import java.io.File;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -60,6 +61,8 @@ public abstract class BatchTestClassGroup extends BaseTestClassGroup {
 			new File(
 				this.gitWorkingDirectory.getWorkingDirectory(),
 				"test.properties"));
+
+		_setTestRelevantChanges();
 	}
 
 	protected String getFirstPropertyValue(
@@ -128,7 +131,33 @@ public abstract class BatchTestClassGroup extends BaseTestClassGroup {
 	protected final String batchName;
 	protected final GitWorkingDirectory gitWorkingDirectory;
 	protected final Properties portalTestProperties;
+	protected boolean testRelevantChanges;
 	protected final String testSuiteName;
+
+	private void _setTestRelevantChanges() {
+		List<String> propertyNames = new ArrayList<>();
+
+		if (testSuiteName != null) {
+			propertyNames.add(
+				JenkinsResultsParserUtil.combine(
+					"test.relevant.changes[", testSuiteName, "]"));
+		}
+
+		propertyNames.add("test.relevant.changes");
+
+		String propertyValue = getFirstPropertyValue(
+			portalTestProperties, propertyNames);
+
+		if (propertyValue != null) {
+			testRelevantChanges = Boolean.parseBoolean(propertyValue);
+
+			return;
+		}
+
+		testRelevantChanges = _DEFAULT_TEST_RELEVANT_CHANGES;
+	}
+
+	private static final boolean _DEFAULT_TEST_RELEVANT_CHANGES = false;
 
 	private final Pattern _propertyNamePattern = Pattern.compile(
 		"[^\\]]+\\[(?<batchName>[^\\]]+)\\](\\[(?<testSuiteName>[^\\]]+)\\])?");
